Remove the link to download Arm GCC
The link to download Arm GCC is intended to be the same for all repos the source files of which are to be compiled. In case the Arm GCC version is to be changed, this link shall be updated in the YML file inside each concerned repo.

Hence, for easier maintainability, this link is removed from the YML files of all these repos and kept only in the YML file of the actions repo.